refactor: Create builders folder in src
