+++ src/builders/MessageBuilder.ts
@@ -1,6 +1,6 @@
-import { Embed } from './types/Embed.js';
-import { EmbedWebhookPayload } from './types/WebhookPayload.js';
-import { formatColor } from './utils.js';
+import { Embed } from '../types/Embed.js';
+import { EmbedWebhookPayload } from '../types/WebhookPayload.js';
+import { formatColor } from '../utils.js';
 
 export class MessageBuilder {
   private content?: string;
